refactored integration tests
